platform--logger/requirements.txt
kafka_python==2.0.2
pymongo==3.11.3
dnspython==2.3.0
azure-storage-blob==12.16.0
